Created a admin namespace
