+++ app/views/admin/activities/_activities_modal.html.haml
@@ -1,4 +1,4 @@
-= form_for [@stage, @step, activity] do |f|
+= form_for [:admin, @stage, @step, activity] do |f|
   .modal.fade{ id: "activityModal#{activity.id}", tabindex: '-1', role: 'dialog', 'aria-labelledby'=> 'activityModalLabel', 'aria-hidden'=> 'true' }
     .modal-dialog
       .modal-content
