+++ app/views/admin/activities/index.html.haml
