+++ app/views/admin/stages/_stages_modal.html.haml
@@ -1,19 +1,19 @@
-= form_for [@stage, step] do |f|
-  .modal.fade{ id: "stepModal#{step.id}", tabindex: '-1', role: 'dialog', 'aria-labelledby'=> 'stepModalLabel', 'aria-hidden'=> 'true' }
+= form_for [:admin, stage] do |f|
+  .modal.fade{ id: "stageModal#{stage.id}", tabindex: '-1', role: 'dialog', 'aria-labelledby'=> 'stageModalLabel', 'aria-hidden'=> 'true' }
     .modal-dialog
       .modal-content
         .modal-header
           %button{ type: 'button', class: 'close', 'data-dismiss'=> 'modal', 'aria-label'=> 'Close' }
             %span{ 'aria-hidden'=> 'true' }
               &times;
-          %h4#myModalLabel.modal-title= t('.step')
+          %h4#myModalLabel.modal-title= t('.stage')
         .modal-body
           .form-group
-            = f.label :name, t('.step_name')
+            = f.label :name, t('.stage_name')
             = f.text_field :name, class: 'form-control'
-          -# We should use this field when it matters the order for each step
+          -# We should use this field when it matters the order for each stage
           -# .form-group
-            = f.label :position, t('.step_position')
+            = f.label :position, t('.stage_position')
             = f.number_field :position, class: 'form-control', 'min' => 1
 
           .form-group
