+++ app/views/admin/steps/_steps_modal.html.haml
@@ -1,19 +1,19 @@
-= form_for stage do |f|
-  .modal.fade{ id: "stageModal#{stage.id}", tabindex: '-1', role: 'dialog', 'aria-labelledby'=> 'stageModalLabel', 'aria-hidden'=> 'true' }
+= form_for [:admin, @stage, step] do |f|
+  .modal.fade{ id: "stepModal#{step.id}", tabindex: '-1', role: 'dialog', 'aria-labelledby'=> 'stepModalLabel', 'aria-hidden'=> 'true' }
     .modal-dialog
       .modal-content
         .modal-header
           %button{ type: 'button', class: 'close', 'data-dismiss'=> 'modal', 'aria-label'=> 'Close' }
             %span{ 'aria-hidden'=> 'true' }
               &times;
-          %h4#myModalLabel.modal-title= t('.stage')
+          %h4#myModalLabel.modal-title= t('.step')
         .modal-body
           .form-group
-            = f.label :name, t('.stage_name')
+            = f.label :name, t('.step_name')
             = f.text_field :name, class: 'form-control'
-          -# We should use this field when it matters the order for each stage
+          -# We should use this field when it matters the order for each step
           -# .form-group
-            = f.label :position, t('.stage_position')
+            = f.label :position, t('.step_position')
             = f.number_field :position, class: 'form-control', 'min' => 1
 
           .form-group
