+++ config/locales/es.yml
@@ -1,58 +1,59 @@
 es:
-  stages:
-    notices:
-      saved_successfully: La etapa fue guardada satisfactoriamente.
-    alerts:
-      save_failed: Hubo un error al guardar los cambios.
-    stages_modal:
-      stage: Etapas
-      stage_name: Nombre de la etapa
-      cancel: Cancelar
-      save: Guardar cambios
-    index:
-      stages: Etapas
-      new_stage: Crear etapa
-      stage_name: Nombre
-      step_count: "# Pasos"
-      no_stages: No se han creado etapas.
-      edit: Editar
-      steps: Pasos
-      actions: Acciones
+  admin:
+    stages:
+      notices:
+        saved_successfully: La etapa fue guardada satisfactoriamente.
+      alerts:
+        save_failed: Hubo un error al guardar los cambios.
+      stages_modal:
+        stage: Etapas
+        stage_name: Nombre de la etapa
+        cancel: Cancelar
+        save: Guardar cambios
+      index:
+        stages: Etapas
+        new_stage: Crear etapa
+        stage_name: Nombre
+        step_count: "# Pasos"
+        no_stages: No se han creado etapas.
+        edit: Editar
+        steps: Pasos
+        actions: Acciones
 
-  steps:
-    notices:
-      saved_successfully: El paso fue guardado satisfactoriamente.
-    alerts:
-      save_failed: Hubo un error al guardar los cambios.
-    steps_modal:
-      step: Pasos
-      step_name: Nombre del paso
-      cancel: Cancelar
-      save: Guardar cambios
-    index:
-      steps: Pasos
-      new_step: Crear paso
-      step_name: Nombre
-      activities_count: "# Actividades"
-      no_steps: No se han creado pasos.
-      edit: Editar
-      activities: Actividades
-      actions: Acciones
+    steps:
+      notices:
+        saved_successfully: El paso fue guardado satisfactoriamente.
+      alerts:
+        save_failed: Hubo un error al guardar los cambios.
+      steps_modal:
+        step: Pasos
+        step_name: Nombre del paso
+        cancel: Cancelar
+        save: Guardar cambios
+      index:
+        steps: Pasos
+        new_step: Crear paso
+        step_name: Nombre
+        activities_count: "# Actividades"
+        no_steps: No se han creado pasos.
+        edit: Editar
+        activities: Actividades
+        actions: Acciones
 
-  activities:
-    notices:
-      saved_successfully: La actividad fue guardada satisfactoriamente.
-    alerts:
-      save_failed: Hubo un error al guardar los cambios.
-    activities_modal:
-      activity: Actividades
-      activity_name: Nombre de la actividad
-      cancel: Cancelar
-      save: Guardar cambios
-    index:
-      activities: Actividades
-      new_activity: Crear actividad
-      activity_name: Nombre
-      no_activities: No se han creado actividades.
-      edit: Editar
-      actions: Acciones
+    activities:
+      notices:
+        saved_successfully: La actividad fue guardada satisfactoriamente.
+      alerts:
+        save_failed: Hubo un error al guardar los cambios.
+      activities_modal:
+        activity: Actividades
+        activity_name: Nombre de la actividad
+        cancel: Cancelar
+        save: Guardar cambios
+      index:
+        activities: Actividades
+        new_activity: Crear actividad
+        activity_name: Nombre
+        no_activities: No se han creado actividades.
+        edit: Editar
+        actions: Acciones
