+++ config/routes.rb
@@ -2,9 +2,11 @@ Rails.application.routes.draw do
   devise_for :users
 
   resources :pages, only: :home
-  resources :stages, only: [:index, :create, :update] do
-    resources :steps, only: [:index, :create, :update] do
-      resources :activities, only: [:index, :create, :update]
+  namespace :admin do
+    resources :stages, only: [:index, :create, :update] do
+      resources :steps, only: [:index, :create, :update] do
+        resources :activities, only: [:index, :create, :update]
+      end
     end
   end
